1.修复BUG选项切换完之后不能选择的BUG  … 2.修复提交SVN时未匹配到BUG与任务信息导致的报错 3.修改登录时禅道地址填错时的报错信息 4.修正站点管理右键编辑和删除菜单的国际化

+++ client/tortoise/TurtleZenTaoLib/IssuesForm.cs
@@ -123,7 +123,7 @@ namespace TurtleZenTaoLib
 
 
             task.name = editForm.getTaskName();
-            task.currentConsumed = Convert.ToString(int.Parse(editForm.getConsumed()) - int.Parse(task.consumed));
+            task.currentConsumed = Convert.ToString(float.Parse(editForm.getConsumed()) - float.Parse(task.consumed));
             task.consumed = editForm.getConsumed();
             task.left = editForm.getLeft();
             task.isDone = editForm.isFinished();
